moved file loaders to base class
